+++ R/add.R
@@ -33,7 +33,7 @@
 #' @export
 add_row <- function(.data, ..., .before = NULL, .after = NULL) {
   if (inherits(.data, "grouped_df")) {
-    stop("Cannot add rows to grouped data frames", call. = FALSE)
+    stop("Can't add rows to grouped data frames", call. = FALSE)
   }
 
   df <- tibble(...)
@@ -41,13 +41,11 @@ add_row <- function(.data, ..., .before = NULL, .after = NULL) {
 
   extra_vars <- setdiff(names(df), names(.data))
   if (length(extra_vars) > 0) {
-    stopc(
-      "This row would add new variables: ", format_n(extra_vars)
-    )
+    stopc(pluralise_msg("Can't add row with new variable(s) ", extra_vars))
   }
 
   missing_vars <- setdiff(names(.data), names(df))
-  df[missing_vars] <- lapply(.data[missing_vars], na_value)
+  df[missing_vars] <- map(.data[missing_vars], na_value)
   df <- df[names(.data)]
 
   pos <- pos_from_before_after(.before, .after, nrow(.data))
@@ -57,10 +55,11 @@ add_row <- function(.data, ..., .before = NULL, .after = NULL) {
 }
 
 na_value <- function(boilerplate) {
-  if (is.list(boilerplate))
+  if (is.list(boilerplate)) {
     list(NULL)
-  else
+  } else {
     NA
+  }
 }
 
 rbind_at <- function(old, new, pos) {
@@ -118,14 +117,19 @@ add_column <- function(.data, ..., .before = NULL, .after = NULL) {
     if (nrow(df) == 1) {
       df <- df[rep(1L, nrow(.data)), ]
     } else {
-      stopc("Expected ", nrow(.data), " rows, got ", nrow(df))
+      stopc(
+        "`.data` must have ", nrow(.data),
+        pluralise_n(" row(s)", nrow(.data)),
+        ", not ", nrow(df)
+      )
     }
   }
 
   extra_vars <- intersect(names(df), names(.data))
   if (length(extra_vars) > 0) {
     stopc(
-      "Columns already in data frame: ", format_n(extra_vars)
+      pluralise_msg("Column(s) ", extra_vars),
+      pluralise(" already exist[s]", extra_vars)
     )
   }
 
@@ -154,17 +158,17 @@ pos_from_before_after_names <- function(before, after, names) {
 }
 
 pos_from_before_after <- function(before, after, len) {
-  if (is.null(before)) {
-    if (is.null(after)) {
+  if (is_null(before)) {
+    if (is_null(after)) {
       len
     } else {
       after
     }
   } else {
-    if (is.null(after)) {
+    if (is_null(after)) {
       before - 1L
     } else {
-      stopc("Can't specify both .before and .after")
+      stopc("Can't specify both `.before` and `.after`")
     }
   }
 }