+++ R/check-names.R
@@ -7,35 +7,34 @@ check_names_df.default <- function(j, ...) {
 
 #' @export
 check_names_df.character <- function(j, x) {
-  check_needs_no_dim(j)
-
-  pos <- safe_match(j, names(x))
-  if(any(is.na(pos))){
-    unknown_names <- j[is.na(pos)]
-    stopc(pluralise_msg("Unknown column(s): ", unknown_names))
-  }
-  pos
+  check_names_before_after.character(j, names(x))
 }
 
 #' @export
 check_names_df.numeric <- function(j, x) {
   check_needs_no_dim(j)
 
-  if (any(is.na(j))) {
+  if (anyNA(j)) {
     stopc("NA column indexes not supported")
   }
 
   non_integer <- (j != trunc(j))
   if (any(non_integer)) {
-    stopc(pluralise_msg("Invalid non-integer column index(es): ", j[non_integer]))
+    stopc(pluralise_msg("Column index(es) must be integer, not ", j[non_integer]))
   }
   neg_too_small <- (j < -length(x))
   if (any(neg_too_small)) {
-    stopc(pluralise_msg("Invalid negative column index(es): ", j[neg_too_small]))
+    stopc(pluralise_msg(
+      paste0("Column index(es) must be at least ", -length(x), " if negative, not "),
+      j[neg_too_small]
+    ))
   }
   pos_too_large <- (j > length(x))
   if (any(pos_too_large)) {
-    stopc(pluralise_msg("Invalid column index(es): ", j[pos_too_large]))
+    stopc(pluralise_msg(
+      paste0("Column index(es) must be at most ", length(x), " if positive, not "),
+      j[pos_too_large]
+    ))
   }
 
   seq_along(x)[j]
@@ -46,9 +45,12 @@ check_names_df.logical <- function(j, x) {
   check_needs_no_dim(j)
 
   if (!(length(j) %in% c(1L, length(x)))) {
-    stopc("Length of logical index vector must be 1 or ", length(x), ", got: ", length(j))
+    stopc(
+      "Length of logical index vector must be 1 or ", length(x),
+      " (the number of rows), not ", length(j)
+    )
   }
-  if (any(is.na(j))) {
+  if (anyNA(j)) {
     stopc("NA column indexes not supported")
   }
   seq_along(x)[j]
@@ -56,10 +58,9 @@ check_names_df.logical <- function(j, x) {
 
 check_needs_no_dim <- function(j) {
   if (needs_dim(j)) {
-    stopc("Unsupported use of matrix or array for column indexing")
+    stopc("Can't use matrix or array for column indexing")
   }
 }
-
 
 # check_names_before_after ------------------------------------------------
 
@@ -75,9 +76,9 @@ check_names_before_after.character <- function(j, names) {
   check_needs_no_dim(j)
 
   pos <- safe_match(j, names)
-  if(any(is.na(pos))){
+  if(anyNA(pos)) {
     unknown_names <- j[is.na(pos)]
-    stopc(pluralise_msg("Unknown column(s): ", unknown_names))
+    stopc(pluralise_msg("Column(s) ", unknown_names), " not found")
   }
   pos
 }